--- 3DGraphicApp/requirements.txt
@@ -1,4 +1,0 @@
-Flask>=2.3.2
-Flask-SQLAlchemy>=3.0.5
-Flask-Migrate>=4.0.5
-gunicorn>=21.2.0
