+++ FunctionSimpler.py
