+++ templates/admin.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="ja">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>管理者としてログイン</title>
+  <style>
+    .admin-form {
+      max-width: 300px;
+      margin: 100px auto;
+      text-align: center;
+    }
+    input, button {
+      display: block;
+      width: 100%;
+      margin-top: 10px;
+    }
+  </style>
+</head>
+<body>
+  <div>
+    <h1>管理者としてログイン</h1>
+    <form method="POST" action="/admin">
+      <input type="password" name="password" placeholder="パスワードを入力" required>
+      <button type="submit">ログイン</button>
+    </form>
+  </div>
+</body>
+</html>
