--- templates/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-  <meta charset="UTF-8">
-  <title>Surface Graphic 3D</title>
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='style.css') }}">
-</head>
-<body>
-  <div class="layout">
-      <form method="POST" action="{{ url_for('index') }}">
-      <input type="text" id="functionInput" name="function_text" placeholder="f(x,y)=" value="{{ function_text or '' }}" />
-    
-      {% if error %}
-        <p class="error">{{ error }}</p>
-      {% endif %}
-    
-      {% if message %}
-        <p>{{ message }}</p>
-      {% endif %}
-
-      <div class="button-panel">
-        <!-- 入力補助ボタン（JSで操作） -->
-        <button type="button" class="input-btn" data-insert="x">x</button>
-        <button type="button" class="input-btn" data-insert="y">y</button>
-        <button type="button" class="input-btn" data-insert="+">+</button>
-        <button type="button" class="input-btn" data-insert="-">-</button>
-        <button type="button" class="input-btn" data-insert="*">×</button>
-        <button type="button" class="input-btn" data-insert="/">÷</button>
-        <button type="button" class="input-btn" data-insert="√">√</button>
-        <button type="button" class="input-btn" data-insert="Power()">^</button>
-        <button type="button" class="input-btn" data-insert="E">e</button>
-
-        <!-- フォームの中に ClickMe ボタンを含める -->
-        <button id="plotBtn" type="submit">ClickMe</button>
-        </div>
-      </form>
-
-      <button id="animateBtn">PLAY ANIMATION</button>
-    </div>
-  </div>
-
-  <canvas id="three-canvas"></canvas>
-
-  <script src="{{ url_for('static', filename='script.js') }}"></script>
-</body>
-</html>
